requirements.txt
# FastAPI and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0

# Database - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Database - MongoDB
motor==3.3.2
pymongo==4.6.1

# Database - Redis
redis==5.0.1
aioredis==2.0.1

# Message Queue & Async Tasks
celery==5.3.4
kombu==5.3.4
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Validation & Serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
orjson==3.9.10

# Reporting
openpyxl==3.1.2
reportlab==4.0.7
jinja2==3.1.2

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.38.0
prometheus-client==0.19.0
